fixing send message

+++ mcast.c
@@ -732,7 +732,7 @@ void initializeAndSendRandomNumber(int moveStartpointer) {
 
 void sendMessage(enum TYPE type, char *dp, int payloadSize) {
 	char message[payloadSize+12];
-	message[0] = type;
+	message[0] = (u_int32_t)type;
 	message[4] = currentSession.machineIndex;
 	message[8] = currentSession.lastDeliveredCounter;
 	memcpy(message + 12, dp, payloadSize);
